backend/requirements.txt
black==23.9.1
celery==5.3.4
Django==4.2.5
django-behaviors==0.5.1
djangorestframework==3.14.0
djoser==2.2.2
drf-spectacular==0.27.1
factory-boy==3.3.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-commas==2.1.0
gunicorn==21.2.0
isort==5.12.0
Pillow==10.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
